Ignore ClusterTerminationTest

+++ ksqldb-rest-app/src/test/java/io/confluent/ksql/api/integration/RestoreCommandTopicMultipleKafkasIntegrationTest.java
@@ -114,7 +114,9 @@ public class RestoreCommandTopicMultipleKafkasIntegrationTest {
   @After
   public void teardown() {
     REST_APP.stop();
-    INTERNAL_TEST_HARNESS.deleteTopics(Collections.singletonList(commandTopic));
+    if (commandTopic != null) {
+      INTERNAL_TEST_HARNESS.deleteTopics(Collections.singletonList(commandTopic));
+    }
   }
 
   private static void writeServerProperties(final Path propertiesFile) throws IOException {
